dev-requirements.txt
-e .
flake8 ~= 3.7
coverage ~= 4.5
aws-cdk.core